+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Update to Bevy 0.18.0-rc.1.
 - Change `TimeKind::default` to `TimeKind::Real`.
+- Add `TimeKind` configuration to `SmoothNudge`.
+- `SmoothNudge::new` no longer `const`.
 
 ## [0.20.1] - 2025-12-29